+++ bundle-es2015/entries/entry-operations.spec.js
@@ -51,6 +51,7 @@ describe('Entry Operations', () => {
     describe('List entries', () => {
         beforeEach(() => {
             setDefaultSpy(global, {
+                pageCount: 1,
                 pageIndex: 0,
                 pageSize: 25,
                 totalCount: 2,
@@ -181,6 +182,7 @@ describe('Entry Operations', () => {
     describe('Search entries', () => {
         beforeEach(() => {
             setDefaultSpy(global, {
+                pageCount: 1,
                 pageIndex: 0,
                 pageSize: 25,
                 totalCount: 2,
@@ -536,6 +538,7 @@ describe('Entry Operations', () => {
     describe('Search entries in IE browser', () => {
         beforeEach(() => {
             setDefaultSpy(global, {
+                pageCount: 1,
                 pageIndex: 0,
                 pageSize: 25,
                 totalCount: 2,